dev-requirements.txt
mock==1.3.0
sphinx
numpydoc
PyYAML>=3.10
ipdb>=0.8
ipython>=2.2.0
coverage==4.0.1
nosexcover==1.0.10
nose==1.3.7
